fix for #3498
(Swap colors on bigendian machines)

I also made it so that the env variable is tried ahead of all
of the other looking.  (That way you can override a system default
if its broken without having to upgrade)

This also makes it so we can remove the OSX ifdef.  So I did.
lukep: If you still want it add it back in but my gut says the less ifdefs
we have the better ;)

Kent


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@6004 954f8c5b-7b00-dc11-b283-0030488c597c