==python scripts== updated version of collada importer
This version includes animation export of underformable objects though IPO channels. It can export animation of transforms and materials. The baked transform options outputs baked transforms for each frame beginning with the lowest transformation frame to the highest frame number for any transformation channel. The animation exporter calculates time values according to the frame number of each key-frame (Bezier-point) depending on the FPS value found in the Scene-Anim settings (defaults to 25).

This version also fixes the texture import on the importer and issues a
warning if the texture could not be loaded and continues to load the
document in any case.

Thanks to Mikael Lagré for the superb update


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@6467 954f8c5b-7b00-dc11-b283-0030488c597c