With the commit last week to have curve bevels nice circular (when no front/back is selected), the UV coordinates for curves should also be corrected. This commit re-uses the same code as for Nurbs, to make sure UV coordinates wrap around nicely.
BUT! I've noticed that Daniel's commit of august in this code actually
broke this UV correction... in his craze to cleanup old code, he missed
the actual functionality. Meaning that in 2.40 and 2.41, "UV orco" texture
coordinates wrap around ugly in Nurbs Surfaces, something that was fixed
in NaN days.

Got no time for tracker now... but I'm sure it's in there! :)


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7023 954f8c5b-7b00-dc11-b283-0030488c597c