-> Stored Selections in Mesh DNA
Stored selections now get saved to mesh library blocks as direct data.
The idea that stored selections are 'erased' when leaving editmode and
switching objects is pretty mysterious for the user. Note that currently
the mselect array in a mesh is not written to file.  Not sure whether
to change this or not.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7191 954f8c5b-7b00-dc11-b283-0030488c597c