Added BPyMesh for mesh python mesh functions. at the moment it only has meshWeight2Dict and dict2MeshWeight   These allow you to deal with vertex weights as a list of dicts which makes scrips short and easy to understand.    (kh_python, perhaps dict access to the python verts could replace this )
Used the above util functions to update mesh_cleanup.
Copied from the source

	'Material Clean', 'Remove unused materials.'
	'VGroups'
	'Group Clean', 'Remove vertex groups that have no verts using them.'
	'Weight Clean', 'Remove zero weighted verts from groups (limit is zero threshold).'
	'Weight Normalize', 'Make the sum total of vertex weights accross vgroups 1.0 for each vertex.'

Normalizing lets you see how much % of the vertex a bone owns just by looking at 1 of the bone weights.

Would be nice to have this functionality in Blender but theres not much room for new buttons in teh vgroup and material area :/


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7211 954f8c5b-7b00-dc11-b283-0030488c597c