=== Transform === Bug #3779: In camera view with center cursor (on camera), it's impossible to use translations.
Initgrabz didn't account for negative zero values (nor near zero values). Works now.

Of course, this is just a safety to give back the appearance of working, since the cursor is on the point of convergence, there's no way to properly calibrate anything.

Word of advice: Always check where the transform center is.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7249 954f8c5b-7b00-dc11-b283-0030488c597c