I'm not sure if I'm allowed to do this, but I'm hoping this is ok... I have been able to compile blender on linux 64 without problems up to a month ago, but a recent change is giving me problems, and probably others as well (on linux 64 at least). But if not, this commit shouldn't change anything. The problem was an added pad variable to RenderData in DNA_scene_types.h, in a somewhat problematic place, since it causes the compiler to pad the struct at the 'ListBase layers' variable, a struct of two pointers, which caused me all sorts of trouble, variables declared after this having incorrect or incorrectly assigned values. This commit fixes this.
In makesdna.c is a useful section which will write a c program called
'padding.c' which compares the actual size to the expected size, which
for the RenderData and Scene struct in this case had a difference of 8.
Also useful was the gcc -Wpadded warning flag.

Hopefully I don't get 'fired' for this... ;)
Now back to fixing the yafray bugs...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7571 954f8c5b-7b00-dc11-b283-0030488c597c