Bugreport #3891
Object Layer Ipos didn't work when the Ipo was moved an Object Action.
Can't get this to work though... this option has been added with a lot
of exception handling already, and for action/nla it's very nasty to
blend/mix/add layer bit values. So; better not support that.

As feedback for users, I've added a notice popup when you try to move
layer-ipos to an action.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7586 954f8c5b-7b00-dc11-b283-0030488c597c