Added support for linked objects from libraries which can have names that are already defined locally, probably does not work fully yet. Added extra 'threads' parameter as requested by Lynx3d.
Optimized drawing of rendered tiles, so that the entire image doesn't have
to be redrawn every time a tile is complete.
The blender code that handles this part was not yet complete and could only
draw (sets of) scanlines. I extended the renderwin_progress() function in
renderwin.c to handle a given subrectangle.
This code needs review!
For the limited test I did it seems to work at least...


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7626 954f8c5b-7b00-dc11-b283-0030488c597c