Bugfix #4535
A proper check for the error "No Camera" on rendering can only be made
after a renderwindow was initialized. The error menu then shows in the
render window, which appears to not work OK in windows ATI (again!).

So; I've moved the test to before the renderwindow is activated, this is
not a test checking on the entire render pipeline (like Composite nodes or
sequence strips). In case more complex setups cannot render, an error is
printed in the console only.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7974 954f8c5b-7b00-dc11-b283-0030488c597c