Added a scanfill wraper to Mathutils, allows you to fill polylines without using Meshes .fill() function. Takes a list of polylines and returns a list of face index triplets.
Added this so using mesh.fill() could be avoided since it requires making an object, linking to the scene and cycling editmode for every NGon imported.
Since this is so close to release, It might be good if ken, willian or stivs takes a look at the function to make sure its ok. - Ran 100's of times for importing lightwave models but would be good to doublecheck.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@7994 954f8c5b-7b00-dc11-b283-0030488c597c