Bugfix #4565
Curves without bevel/extrude don't react to buttons to fill front/back.
That is coded that way, so added this info to the tooltip

Bug report #4566

Creases drawing became very awkward in commit in august last year. Flipped
the drawing order so the fat lines - denoting creases - are drawn first.
That way the value of crease is visible, as well as edge selection state.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8009 954f8c5b-7b00-dc11-b283-0030488c597c