== FFMPEG ==
Rewrote the video seek code again. This time it is a _lot_ simpler and a lot
better.

For those interested: instead of letting ffmpeg choose the right final
position, we let it preseek (maybe to I-frame in the rare case, that is
implemented) and then scan for the correct frame on DTS basis by hand.

We therefor only have to seek once and remove all those frame-jumping
jerkiness, people noticed in the same step.

Tested with msmpeg, hd-mpeg2 and DV.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8100 954f8c5b-7b00-dc11-b283-0030488c597c