Scripts: - Updating Blender Lip Synchro to the newest version sent by the author, Benoit Foucque (Dienben). Thanks!
I'm committing a version with a few minor (tested) changes to make it
easier and to fix a small problem, the author received the changes a few
days ago. He would try to send another update implementing a couple
suggestions, but it may not come in time, so better have this. Anyway,
interested users can check his blog for test files and more information:

http://blenderlipsynchro.blogspot.com/


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8150 954f8c5b-7b00-dc11-b283-0030488c597c