-> Bugfix (again)
Didn't realize that LMB+ALT was used for middle mouse emulation. Changed
the vertex snap modifier key for knife tool to 'CTRL' and fixed problems
with the drawing of the header instructions.

Also changed color of the knife tool to match that of loopcut tool. This
is easier to see and more consistent from a UI standpoint as well.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8307 954f8c5b-7b00-dc11-b283-0030488c597c