Bugfix #4843 revisited
This is a new incarnation of compatible_eul(), a function called:

void Mat3ToCompatibleEul(float mat[][3], float *eul, float *oldrot)

It uses the two euler extraction methods as added by Brecht a while ago,
and checks for compatibility each, and then picks the best of the two
based on minimal difference with 'oldrot'.

Gives for rotation key inserting a much higher hitrate for OK eulers.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8313 954f8c5b-7b00-dc11-b283-0030488c597c