Fix for a bug where the Array modifier would freeze if it was set to "Fit To Curve Length" or "Fixed Count" and the base mesh was scaled to 0 in edit mode (could also happen while entering a numerical scale value like 0.25).
The problem was that the dist value could be almost 0, leading to a
ridiculously large duplicate count which would then cause memory allocation
to fail and the array modifier to get stuck in an almost infinite loop trying
to calculate the offset of the final copy. This commit fixes the problem
by checking that dist is greater than FLT_EPSILON before continuing.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8418 954f8c5b-7b00-dc11-b283-0030488c597c