Patch #5069, by Juho Vepsäläinen (bebraw)
Enable to use Xkey or DEL for deleting objects in outliner.

Note: there's reasons to be a bit reluctant with extensions like this.
Mostly related to the fact that outliner operations is completely unfinished
still. We need to spend time on thinking over how it should evolve, and
based on that the proper hotkeys and menus can be added as well.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@8788 954f8c5b-7b00-dc11-b283-0030488c597c