Proxy Objects revisited!
The first incarnation assumed that proxies were local objects per
definition. Unfortunately that makes it impossible to - for example -
reference-link an entire Scene with proxies, to be used as a special
character set.

This commit makes the proxy implementation also a bit more clear.

Related work: the scene-sets were not executed fully or correctly for
the dependency graph. That happens now (in 3d view) as well.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@9124 954f8c5b-7b00-dc11-b283-0030488c597c